Move class_size_predictor.py to src directory and change relevant imports
